opencl with cost

+++ src/pomelo_avx2_plug.c
@@ -79,7 +79,7 @@
         S[index_local] = ADD256(S[index_local],  SHIFTL256(S[i0],2));               \
         S[i0]          = ADD256(S[i0],SHIFTL256(S[index_global],1));                \
         S[index_global]= ADD256(S[index_global], SHIFTL256(S[i0],3));               \
-        random_number  = ((unsigned long long*)S)[(i3 << 2)];                       \
+        random_number  = ((uint64_t*)S)[(i3 << 2)];                       \
     }                                                                               \
 }
 
@@ -106,11 +106,11 @@ static void aligned_free(void *p)
 int POMELO_AVX2(void *out, size_t outlen, const void *in, size_t inlen,
     const void *salt, size_t saltlen, unsigned int t_cost, unsigned int m_cost)
 {
-	unsigned long long i, j, k;
-	unsigned long long i0, i1, i2, i3, i4;
+	uint64_t i, j, k;
+	uint64_t i0, i1, i2, i3, i4;
 	__m256i *S;
-	unsigned long long random_number, index_global, index_local;
-	unsigned long long state_size, mask;
+	uint64_t random_number, index_global, index_local;
+	uint64_t state_size, mask;
 
 	//check the size of password, salt, and output. Password at most 256 bytes; salt at most 64 bytes; output at most 256 bytes.  
 	if (inlen > 256 || saltlen > 64 || outlen > 256 || inlen < 0 ||
